azure_functions/requirements.txt
azure-functions
azure-identity>=1.12.0
azure-digitaltwins-core>=1.2.0
